feat(x86_64): Full 64-bit Long Mode Edition - 4-Level Paging (PML4), 64-bit GDT/IDT (iretq), SysV x86_64 ABI, 64-bit PMM & 16MB Heap, and 64-bit GUI Desktop

+++ src/include/heap.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-void heap_init(uint32_t start_addr, size_t size);
+void heap_init(uint64_t start_addr, size_t size);
 void* kmalloc(size_t size);
 void* kcalloc(size_t num, size_t size);
 void* krealloc(void* ptr, size_t size);
