+++ src/include/pmm.h
@@ -5,7 +5,7 @@
 
 #define PAGE_SIZE 4096
 
-void pmm_init(uint32_t mem_size_bytes);
+void pmm_init(uint64_t mem_size_bytes);
 void* pmm_alloc_page(void);
 void pmm_free_page(void* ptr);
 size_t pmm_get_free_pages(void);
